Navigation, config & background
The navigation is set up but currently it's not a livewire component. This is because I don't think it needs to be and to save on requests it's better not to make it one for now at least.

Edited the tailwind config for the container stuff.

And added the background image